move all source files to `src/`
