+++ src/editor.c
