+++ src/elf.c
