+++ src/hdl_rpc.c
