+++ src/jpgviewer.c
