+++ src/makeicon.c
