+++ src/timer.c
